Move attachment support to SimpleEmailBuilder, add inline support to it
