+++ src/main/kotlin/pt/up/fe/ni/website/backend/email/BaseEmailBuilder.kt
@@ -1,12 +1,7 @@
 package pt.up.fe.ni.website.backend.email
 
-import jakarta.activation.DataSource
-import jakarta.activation.FileDataSource
-import jakarta.activation.URLDataSource
 import jakarta.validation.Valid
 import jakarta.validation.constraints.Email
-import java.io.File
-import java.net.URL
 import org.springframework.mail.javamail.MimeMessageHelper
 import pt.up.fe.ni.website.backend.config.email.EmailConfigProperties
 import pt.up.fe.ni.website.backend.model.Account
@@ -17,7 +12,6 @@ abstract class BaseEmailBuilder : EmailBuilder {
     private var to: MutableSet<String> = mutableSetOf()
     private var cc: MutableSet<String> = mutableSetOf()
     private var bcc: MutableSet<String> = mutableSetOf()
-    private var attachments: MutableList<Attachment> = mutableListOf()
 
     fun from(@Email email: String, personal: String = email) = apply {
         from = email
@@ -48,30 +42,11 @@ abstract class BaseEmailBuilder : EmailBuilder {
         bcc.addAll(users.map { it.email })
     }
 
-    fun attach(name: String, content: DataSource) = apply {
-        attachments.add(Attachment(name, content))
-    }
-
-    fun attach(name: String, content: File) = apply {
-        attachments.add(Attachment(name, FileDataSource(content)))
-    }
-
-    fun attach(name: String, path: String) = apply {
-        attachments.add(Attachment(name, URLDataSource(URL(path))))
-    }
-
     override fun build(helper: MimeMessageHelper, emailConfigProperties: EmailConfigProperties) {
         helper.setFrom(from ?: emailConfigProperties.from, fromPersonal ?: emailConfigProperties.fromPersonal)
 
         to.forEach(helper::setTo)
         cc.forEach(helper::setCc)
         bcc.forEach(helper::setBcc)
-
-        attachments.forEach { helper.addAttachment(it.name, it.content) }
     }
-
-    protected data class Attachment(
-        val name: String,
-        val content: DataSource
-    )
 }
