+++ src/main/kotlin/pt/up/fe/ni/website/backend/email/SimpleEmailBuilder.kt
@@ -1,5 +1,10 @@
 package pt.up.fe.ni.website.backend.email
 
+import jakarta.activation.DataSource
+import jakarta.activation.FileDataSource
+import jakarta.activation.URLDataSource
+import java.io.File
+import java.net.URL
 import org.springframework.mail.javamail.MimeMessageHelper
 import pt.up.fe.ni.website.backend.config.email.EmailConfigProperties
 
@@ -7,6 +12,8 @@ class SimpleEmailBuilder : BaseEmailBuilder() {
     private var text: String? = null
     private var html: String? = null
     private var subject: String? = null
+    private var attachments: MutableList<EmailFile> = mutableListOf()
+    private var inlines: MutableList<EmailFile> = mutableListOf()
 
     fun text(text: String) = apply {
         this.text = text
@@ -18,6 +25,30 @@ class SimpleEmailBuilder : BaseEmailBuilder() {
 
     fun subject(subject: String) = apply {
         this.subject = subject
+    }
+
+    fun attach(name: String, content: DataSource) = apply {
+        attachments.add(EmailFile(name, content))
+    }
+
+    fun attach(name: String, content: File) = apply {
+        attachments.add(EmailFile(name, FileDataSource(content)))
+    }
+
+    fun attach(name: String, path: String) = apply {
+        attachments.add(EmailFile(name, URLDataSource(URL(path))))
+    }
+
+    fun inline(name: String, content: DataSource) = apply {
+        inlines.add(EmailFile(name, content))
+    }
+
+    fun inline(name: String, content: File) = apply {
+        inlines.add(EmailFile(name, FileDataSource(content)))
+    }
+
+    fun inline(name: String, path: String) = apply {
+        inlines.add(EmailFile(name, URLDataSource(URL(path))))
     }
 
     override fun build(helper: MimeMessageHelper, emailConfigProperties: EmailConfigProperties) {
@@ -34,5 +65,13 @@ class SimpleEmailBuilder : BaseEmailBuilder() {
         if (subject != null) {
             helper.setSubject(subject!!)
         }
+
+        attachments.forEach { helper.addAttachment(it.name, it.content) }
+        inlines.forEach { helper.addInline(it.name, it.content) }
     }
+
+    private data class EmailFile(
+        val name: String,
+        val content: DataSource
+    )
 }
